# degensample-backend/api_docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "defi-relay-backend",
    "description": "A protocol for web3 payments.",
    "license": {
      "name": "MIT"
    },
    "version": "0.2.0"
  },
  "paths": {
    "/api/apikey/create": {
      "post": {
        "tags": [
          "controllers::api_key_controller"
        ],
        "operationId": "create_api_key",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Creates an api key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse_CreateApiKeyInput"
                }
              }
            }
          }
        }
      }
    },
    "/api/apikey/delete": {
      "post": {
        "tags": [
          "controllers::api_key_controller"
        ],
        "operationId": "delete_api_key",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteApiKeyInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Deletes an API key by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse_DeleteApiKeyOutput"
                }
              }
            }
          }
        }
      }
    },
    "/api/apikey/list": {
      "get": {
        "tags": [
          "controllers::api_key_controller"
        ],
        "operationId": "list_api_keys",
        "parameters": [
          {
            "name": "session_token",
            "in": "query",
            "description": "User's session token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "wallet_public_address",
            "in": "query",
            "description": "User's wallet address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lists all API keys for a user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse_Vec_ListApiKeysOutput"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/payments/find_by_invoice_uuid": {
      "get": {
        "tags": [
          "controllers::payments_controller"
        ],
        "operationId": "find_payment_by_invoice_uuid",
        "parameters": [
          {
            "name": "invoice_uuid",
            "in": "path",
            "description": "UUID of the invoice to find payment for",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/DomainBytes32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Payment not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments/list": {
      "post": {
        "tags": [
          "controllers::payments_controller"
        ],
        "operationId": "list_payments",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListPaymentsQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "List of payments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse_String"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse_String"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthResponse_CreateApiKeyInput": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "session_token",
              "wallet_public_address"
            ],
            "properties": {
              "name": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "session_token": {
                "type": "string"
              },
              "wallet_public_address": {
                "type": "string"
              }
            }
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "AuthResponse_DeleteApiKeyOutput": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "deleted"
            ],
            "properties": {
              "deleted": {
                "type": "boolean"
              }
            }
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "AuthResponse_String": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "data": {
            "type": "string"
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "AuthResponse_Vec_ListApiKeysOutput": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "api_key_id",
                "api_key",
                "created_at"
              ],
              "properties": {
                "api_key": {
                  "type": "string"
                },
                "api_key_id": {
                  "type": "integer",
                  "format": "int32"
                },
                "created_at": {
                  "type": "integer",
                  "format": "int64"
                },
                "name": {
                  "type": [
                    "string",
                    "null"
                  ]
                }
              }
            }
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ColumnSortDir": {
        "type": "string",
        "enum": [
          "Asc",
          "Desc"
        ]
      },
      "CreateApiKeyInput": {
        "type": "object",
        "required": [
          "session_token",
          "wallet_public_address"
        ],
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "session_token": {
            "type": "string"
          },
          "wallet_public_address": {
            "type": "string"
          }
        }
      },
...
      "DeleteApiKeyInput": {
        "type": "object",
        "required": [
          "api_key"
        ],
        "properties": {
          "api_key": {
            "type": "string"
          }
        }
      },
...
      "DomainBytes32": {
        "type": "string",
        "description": "32-byte value represented as a hex string"
      },
...
      "ListPaymentsQuery": {
        "type": "object",
        "required": [
          "session_token"
        ],
        "properties": {
          "chain_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "pagination": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PaginationData"
              }
            ]
          },
          "session_token": {
            "type": "string"
          }
        }
      },
...
      "PaginationData": {
        "type": "object",
        "properties": {
          "page": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "page_size": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "sort_by": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TinySafeString"
              }
            ]
          },
          "sort_dir": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ColumnSortDir"
              }
            ]
          }
        }
      },
...
      "TinySafeString": {
        "type": "string"
      },
...
    }
  }
}